fixing a few man pages
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/genefilter@24691 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/nd.c
@@ -26,7 +26,7 @@
 #if defined(USING_R) /*( R-specific stuff */
 
 #define  S_CDECL
-typedef int RSInt;
+typedef long RSInt;
 
 #ifdef HAVE_CONFIG_H
 # include <config.h>
